requirements.txt
openai==1.3.8 ; python_version >= "3.10" and python_version < "4.0"
tenacity==8.2.2 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.24.0.post1
requests
fastapi==0.105.0
python-dotenv
azure-common==1.1.28
azure-core==1.29.5
azure-cosmos==4.5.0
azure-identity==1.14.0
azure-search-documents==11.4.0
azure-storage-blob==12.17.0
langdetect==1.0.9